+++ docs/en/changes/changes.md
@@ -70,6 +70,8 @@
 * Add type descriptor when converting Envoy logs to JSON for persistence, to avoid conversion error.
 * Bseline: Support query baseline with MQE and use in the Alarm Rule.
 * Bump up netty to 4.11.118 to fix CVE-2025-24970.
+* Add `Get Alarm Runtime Status` API.
+* Add `lock` when query the Alarm metrics window values.
 
 #### UI